requirements.txt
torch>=2.4.0
transformers>=4.51.0
trl>=0.14.0
peft>=0.14.0
accelerate>=1.2.0
bitsandbytes>=0.45.0
datasets>=3.0.0
python-chess>=1.999
stockfish
wandb
langdetect
zstandard
requests
